Refactor realtime api interfaces into realtime pkg
